refactor: ♻️ 删除TSPluginAdaptor文件，将其内容放入constants.ts和fileController.ts中
